rename bluecellulab.cell.cell.Cell -> bluecellulab.cell.core.Cell
