+++ bluecellulab/cell/__init__.py
@@ -1,5 +1,5 @@
 """The cell module representing the NEURON cell and operations on it."""
 
-from .cell import Cell
+from .core import Cell
 from .cell_dict import CellDict
 from .sonata_proxy import SonataProxy
