+++ bluecellulab/cell/core.py
